look for a space after "tip " before responding
If your bot is named @tipbot, then often people will be confused when
they type "tipbot balance" and get back "I didn't understand that command".
That was happening because "tipbot balance" matches /tip/ but is not a
valid tipping command. Requiring the space means "tipbot balance" will
just be ignored.